Add orange.txt and update file reading in main.rs
A new file, orange.txt, was created with a list of fruit names. Moreover, the manner in which files are read and processed in main.rs has now been updated. The major change in file processing is string conversion of the content and a debug print of it. The process includes reading, updating the buffer, and hashing the entire buffer instead of just a segment.