Add 'commit-tree' command and refactor 'hash-object' command
Added the functionality for 'commit-tree' in the CLI, allowing users to create a new commit object. Also refactored 'hash-object' code by moving the file reading and writing ownership to the 'write_object' function. Lastly, enhanced 'write_object' to take in the kind of object and requirement to write to file as arguments.